Skip invalid usernames

+++ src/notifications.py
@@ -1,4 +1,5 @@
 import discord_logging
+import re
 
 import utils
 import static
@@ -22,29 +23,35 @@ def send_reminders(reddit, database):
 			reminders_sent += 1
 			counters.notifications.inc()
 			counters.queue.dec()
-			log.info(
-				f"{reminders_sent}/{len(reminders)}/{count_reminders}: Sending reminder to u/{reminder.user.name} : "
-				f"{reminder.id} : {utils.get_datetime_string(reminder.target_date)}")
-			bldr = utils.get_footer(reminder.render_notification())
-			result = reddit.send_message(reminder.user.name, "RemindMeBot Here!", ''.join(bldr))
-			if result in [ReturnType.INVALID_USER, ReturnType.USER_DOESNT_EXIST]:
-				log.info(f"User doesn't exist: u/{reminder.user.name}")
-			if result in [ReturnType.NOT_WHITELISTED_BY_USER_MESSAGE]:
-				log.info(f"User blocked notification message: u/{reminder.user.name}")
-
-			if reminder.recurrence is not None:
-				if reminder.user.recurring_sent > static.RECURRING_LIMIT:
-					log.info(f"User u/{reminder.user.name} hit their recurring limit, deleting reminder {reminder.id}")
-					database.delete_reminder(reminder)
-				else:
-					new_target_date = utils.parse_time(reminder.recurrence, reminder.target_date + timedelta(seconds=1), reminder.user.timezone) - timedelta(seconds=1)
-					log.info(f"{reminder.id} recurring from {utils.get_datetime_string(reminder.target_date)} to "
-							 f"{utils.get_datetime_string(new_target_date)}")
-					reminder.target_date = new_target_date
-					reminder.user.recurring_sent += 1
-			else:
+			if re.search(r"[^\w_-]", reminder.user.name):
+				log.warning(f"Can't send reminder, invalid username: u/{reminder.user.name} : {reminder.id} : {utils.get_datetime_string(reminder.target_date)}")
 				log.debug(f"{reminder.id} deleted")
 				database.delete_reminder(reminder)
+
+			else:
+				log.info(
+					f"{reminders_sent}/{len(reminders)}/{count_reminders}: Sending reminder to u/{reminder.user.name} : "
+					f"{reminder.id} : {utils.get_datetime_string(reminder.target_date)}")
+				bldr = utils.get_footer(reminder.render_notification())
+				result = reddit.send_message(reminder.user.name, "RemindMeBot Here!", ''.join(bldr))
+				if result in [ReturnType.INVALID_USER, ReturnType.USER_DOESNT_EXIST]:
+					log.info(f"User doesn't exist: u/{reminder.user.name}")
+				if result in [ReturnType.NOT_WHITELISTED_BY_USER_MESSAGE]:
+					log.info(f"User blocked notification message: u/{reminder.user.name}")
+
+				if reminder.recurrence is not None:
+					if reminder.user.recurring_sent > static.RECURRING_LIMIT:
+						log.info(f"User u/{reminder.user.name} hit their recurring limit, deleting reminder {reminder.id}")
+						database.delete_reminder(reminder)
+					else:
+						new_target_date = utils.parse_time(reminder.recurrence, reminder.target_date + timedelta(seconds=1), reminder.user.timezone) - timedelta(seconds=1)
+						log.info(f"{reminder.id} recurring from {utils.get_datetime_string(reminder.target_date)} to "
+								 f"{utils.get_datetime_string(new_target_date)}")
+						reminder.target_date = new_target_date
+						reminder.user.recurring_sent += 1
+				else:
+					log.debug(f"{reminder.id} deleted")
+					database.delete_reminder(reminder)
 
 		database.commit()
